+++ x/pocketcore/keeper/nodes.go
@@ -46,7 +46,7 @@ func (k Keeper) GetSelfNode(ctx sdk.Ctx) (node exported.ValidatorI, er sdk.Error
 }
 
 // "AwardCoinsForRelays" - Award coins to nodes for relays completed using the nodes keeper
-func (k Keeper) AwardCoinsForRelays(ctx sdk.Ctx, relays int64, toAddr sdk.Address) sdk.Int {
+func (k Keeper) AwardCoinsForRelays(ctx sdk.Ctx, relays int64, toAddr sdk.Address) sdk.BigInt {
 	return k.posKeeper.RewardForRelays(ctx, sdk.NewInt(relays), toAddr)
 }
